app/requirements.txt
fastapi
pandas
joblib
nltk
python-dotenv
requests
uvicorn
